shoppingcart/wishlist pages, wishlist edit size TODO

+++ Backend/springboot-jwt-auth/src/main/java/com/tericcabrel/authapi/services/WishlistService.java
@@ -64,7 +64,7 @@ public class WishlistService {
         return new WishlistDto(userWishlist);
     }
 
-    public WishlistDto getUserWishlist(Long id) {
+    public WishlistDto getUserWishlist(Integer id) {
         return new WishlistDto(wishlistRepository.findByUserId(id));
     }
